added faster tool module loading
