+++ src/js/pages/app.js
@@ -23,7 +23,6 @@ import * as OnlineModeActions from '../actions/OnlineModeActions';
 import * as MigrationActions from '../actions/MigrationActions';
 import * as SettingsMigrationActions from '../actions/SettingsMigrationActions';
 import { loadLocalization, APP_LOCALE_SETTING } from '../actions/LocaleActions';
-import { getToolsMetadatas } from '../actions/ToolsMetadataActions';
 import {getLocaleLoaded, getSetting} from '../selectors';
 
 import packageJson from '../../../package.json';
@@ -48,8 +47,7 @@ class Main extends Component {
     const {
       migrateResourcesFolder,
       migrateToolsSettings,
-      getAnchorTags,
-      getToolsMetadatas
+      getAnchorTags
     } = this.props;
 
     const tcResourcesPath = path.join(ospath.home(), 'translationCore', 'resources');
@@ -60,7 +58,6 @@ class Main extends Component {
     // migration logic for toolsSettings in settings.json
     migrateToolsSettings();
     getAnchorTags();
-    // getToolsMetadatas();
   }
 
   render() {
@@ -100,8 +97,7 @@ Main.propTypes = {
   migrateToolsSettings: PropTypes.func.isRequired,
   getAnchorTags: PropTypes.func.isRequired,
   isLocaleLoaded: PropTypes.bool,
-  appLanguage: PropTypes.any,
-  getToolsMetadatas: PropTypes.func.isRequired
+  appLanguage: PropTypes.any
 };
 
 const mapStateToProps = state => {
@@ -115,8 +111,7 @@ const mapDispatchToProps = {
   getAnchorTags: OnlineModeActions.getAnchorTags,
   migrateToolsSettings: SettingsMigrationActions.migrateToolsSettings,
   migrateResourcesFolder: MigrationActions.migrateResourcesFolder,
-  loadLocalization,
-  getToolsMetadatas
+  loadLocalization
 };
 
 export default connect(
